feat: Account, Btc 테이블 추가 (#6) 테스트 실패
- 두 테이블 모두, 회원가입 시에 자동으로 생성
- 현재가 및 지정가 주문과 동일하게 실행
- 테스트 실패

+++ src/main/java/com/example/mockbit/order/application/OrderService.java
@@ -1,5 +1,6 @@
 package com.example.mockbit.order.application;
 
+import com.example.mockbit.account.application.AccountService;
 import com.example.mockbit.common.infrastructure.redis.RedisService;
 import com.example.mockbit.order.domain.Order;
 import lombok.RequiredArgsConstructor;
@@ -7,6 +8,7 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
+import java.math.BigDecimal;
 import java.time.Instant;
 import java.util.List;
 import java.util.Optional;
@@ -19,14 +21,16 @@ import java.util.stream.Collectors;
 public class OrderService {
 
     private final RedisService redisService;
+    private final AccountService accountService;
 
     private static final String REDIS_ORDER_KEY = "Orders:%s:%d";
 
     @Transactional
     public Order saveOrder(Long userId, String price, String btcPrice, String orderPrice, int leverage,
                            String position, String sellOrBuy) {
+        String redisKey = String.format(REDIS_ORDER_KEY, price, userId);
         Order order = Order.builder()
-                .id(String.format("%s:%d", price, userId))
+                .id(redisKey)
                 .price(price)
                 .userId(userId)
                 .orderedAt(String.valueOf(Instant.now()))
@@ -37,8 +41,8 @@ public class OrderService {
                 .sellOrBuy(sellOrBuy)
                 .build();
 
-        String redisKey = String.format(REDIS_ORDER_KEY, price, userId);
         redisService.saveData(redisKey, order);
+        accountService.processOrder(userId, BigDecimal.valueOf(Long.parseLong(order.getOrderPrice())));
         log.info("지정가 주문이 등록되었습니다. - User: {}, Price: {}", userId, price);
 
         return order;
@@ -61,6 +65,20 @@ public class OrderService {
 
     @Transactional
     public void deleteOrderById(String id) {
+        Order order = (Order) redisService.getData(id);
+        if (order == null) {
+            throw new RuntimeException("주문을 찾을 수 없습니다.");
+        }
+
         redisService.deleteData(id);
+
+        String[] parts = id.replaceFirst("Orders:", "").split(":");
+
+        if (parts.length < 2) {
+            throw new RuntimeException("주문 ID 형식이 올바르지 않습니다.");
+        }
+
+        Long userId = Long.valueOf(parts[1]);
+        accountService.cancelOrder(userId, BigDecimal.valueOf(Long.parseLong(order.getOrderPrice())));
     }
 }
